requirements.txt
daemon
Django<1.9
django-bootstrap3
django-revproxy
django-tables2<1.17
django-dbbackup<3.0
gitdb
GitPython
psutil
Pygments
pytz
requests
six
smmap
urllib3<1.17
ipy
elasticsearch
django-ipware
djangorestframework                                                                                                                                                                                               
markdown                                                                                                                                                                                                          
django-filter
